tracing: Move LogValueSet into __macro_support
